Added Harvester Script
I added a script to the harvester that allows it to autonomously harvest ore from ore piles after initial direction from the human player. Ore now also disappears after when it is used up. For now, the only reward for the player when ore is deposited is a console log.