Added an RTS-esque Camera
This camera looks for objects with a new "Ground" tag in the middle of the screen. You arne't allowed to move the camera into a position that does not have the ground in the middle of the screen. The camera height automatically adjusts based on the y coordinate of the ground object in the middle of the screen. If the object in the middle of the screen is a unit/structure, the script looks nearby.