Implemented Advanced Camera Functionality
Implemented panning and adjustable relative height to ground for the camera. Panning is done via right click, which is an overload of the move/attack command, but the TFM camera script has an algorithm to determine which is desired. Additionally, a timeout of one second for side-scrolling after panning has been implemented. That is useful for since your cursor may exceed the game's viewing area when panning.