Adding Slots Tag, Changing Colors, Fixing Pause Menu
Changing the text colors on the HUD to ensure that its easily readible. Additionally making all of the slots tagged under the PlayerSlots tag. The pause menu was having some visual issues due to alignment of the buttons. This was adjusted to ensure that the UI does not glitch weirdly when paused.