Enemy AI Enhancements
This commit includes three AI improvements: the Enemy now attempts to build up to having threee harvesters. Enemy units are less likely to fight for moving to the center of a rally point now. There is now a public variable for adjusting the search and destroy distance of the enemy AI per unit.